+++ lib/project_types/script/layers/infrastructure/push_package_repository.rb
@@ -7,23 +7,6 @@ module Script
         include SmartProperties
         property! :ctx, accepts: ShopifyCLI::Context
 
-        def create_push_package(script_project:, script_content:, metadata:, library:)
-          build_file_path = file_path(script_project.id)
-          write_to_path(build_file_path, script_content)
-
-          Domain::PushPackage.new(
-            id: build_file_path,
-            uuid: script_project.uuid,
-            extension_point_type: script_project.extension_point_type,
-            title: script_project.title,
-            description: script_project.description,
-            script_content: script_content,
-            metadata: metadata,
-            script_config: script_project.script_config,
-            library: library
-          )
-        end
-
         def get_push_package(script_project:, metadata:, library:)
           build_file_path = file_path(script_project.id)
           raise Domain::Errors::PushPackageNotFoundError unless ctx.file_exist?(build_file_path)
@@ -44,13 +27,8 @@ module Script
 
         private
 
-        def write_to_path(path, content)
-          ctx.mkdir_p(File.dirname(path))
-          ctx.binwrite(path, content)
-        end
-
         def file_path(path_to_script)
-          "#{path_to_script}/build/script.wasm"
+          "#{path_to_script}/build/index.wasm"
         end
       end
     end