+++ lib/shopify_cli/theme/syncer.rb
@@ -7,15 +7,15 @@ require "forwardable"
 require_relative "syncer/error_reporter"
 require_relative "syncer/standard_reporter"
 require_relative "syncer/operation"
+require_relative "theme_admin_api"
 
 module ShopifyCLI
   module Theme
     class Syncer
       extend Forwardable
 
-      API_VERSION = "unstable"
-
       attr_reader :checksums
+      attr_reader :checksums_mutex
       attr_accessor :include_filter
       attr_accessor :ignore_filter
 
@@ -39,11 +39,18 @@ module ShopifyCLI
         # Mutex used to pause all threads when backing-off when hitting API rate limits
         @backoff_mutex = Mutex.new
 
+        # Mutex used to coordinate changes in the checksums (shared accross all threads)
+        @checksums_mutex = Mutex.new
+
         # Latest theme assets checksums. Updated on each upload.
         @checksums = {}
 
         # Checksums of assets with errors.
         @error_checksums = []
+      end
+
+      def api_client
+        @api_client ||= ThemeAdminAPI.new(@ctx, @theme.shop)
       end
 
       def lock_io!
@@ -96,11 +103,8 @@ module ShopifyCLI
       end
 
       def fetch_checksums!
-        _status, response = ShopifyCLI::AdminAPI.rest_request(
-          @ctx,
-          shop: @theme.shop,
-          path: "themes/#{@theme.id}/assets.json",
-          api_version: API_VERSION,
+        _status, response = api_client.get(
+          path: "themes/#{@theme.id}/assets.json"
         )
         update_checksums(response)
       end
@@ -239,12 +243,8 @@ module ShopifyCLI
           asset[:attachment] = Base64.encode64(file.read)
         end
 
-        _status, body, response = ShopifyCLI::AdminAPI.rest_request(
-          @ctx,
-          shop: @theme.shop,
+        _status, body, response = api_client.put(
           path: "themes/#{@theme.id}/assets.json",
-          method: "PUT",
-          api_version: API_VERSION,
           body: JSON.generate(asset: asset)
         )
 
@@ -276,12 +276,8 @@ module ShopifyCLI
       end
 
       def get(file)
-        _status, body, response = ShopifyCLI::AdminAPI.rest_request(
-          @ctx,
-          shop: @theme.shop,
+        _status, body, response = api_client.get(
           path: "themes/#{@theme.id}/assets.json",
-          method: "GET",
-          api_version: API_VERSION,
           query: URI.encode_www_form("asset[key]" => file.relative_path.to_s),
         )
 
@@ -298,12 +294,8 @@ module ShopifyCLI
       end
 
       def delete(file)
-        _status, _body, response = ShopifyCLI::AdminAPI.rest_request(
-          @ctx,
-          shop: @theme.shop,
+        _status, _body, response = api_client.delete(
           path: "themes/#{@theme.id}/assets.json",
-          method: "DELETE",
-          api_version: API_VERSION,
           body: JSON.generate(asset: {
             key: file.relative_path.to_s,
           })
@@ -314,14 +306,17 @@ module ShopifyCLI
 
       def update_checksums(api_response)
         api_response.values.flatten.each do |asset|
-          if asset["key"]
+          next unless asset["key"]
+          checksums_mutex.synchronize do
             @checksums[asset["key"]] = asset["checksum"]
           end
         end
         # Generate .liquid asset files are reported twice in checksum:
         # once of generated, once for .liquid. We only keep the .liquid, that's the one we have
         # on disk.
-        @checksums.reject! { |key, _| @checksums.key?("#{key}.liquid") }
+        checksums_mutex.synchronize do
+          @checksums.reject! { |key, _| @checksums.key?("#{key}.liquid") }
+        end
       end
 
       def file_has_changed?(file)